backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
scikit-learn
anthropic
pandas
numpy
pytest
httpx
google-generativeai
